Implement most of composition and Kan filling (Kan fills aren't parsed yet).
This seems to work consistently, but is currently not very useful because we don't normalise interval vars enough in the presence of face restrictions.
I'm not sure what the best method for doing this is. I feel like I'm duplicating a lot of functionality between face formulae and the context however I do it. Maybe needs some rethinking..